Refactor the main generator class to allow unit testing of the production logic
